Refactor CalendarConverterBox for backwards compatibility
- Re-exported the main component from the converter package to maintain compatibility with existing code.
- Removed the previous implementation of the CalendarConverterBox, simplifying the structure while preserving functionality.
